mindbloom-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-generativeai==0.8.5
speechrecognition==3.10.0
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1 